docker/ga3-microservices/attachment-service/app/serializers.py
from rest_framework import serializers
from .models import Attachments

class AttachmentsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Attachments
        fields = '__all__'